Remove Go SSE server and EventSource connection (#4020)
