+++ cmd/publisher/commands/ui.go
@@ -3,13 +3,14 @@ package commands
 // Copyright (C) 2023 by Posit Software, PBC.
 
 import (
+	"os"
+
 	"github.com/posit-dev/publisher/internal/accounts"
 	"github.com/posit-dev/publisher/internal/cli_types"
 	"github.com/posit-dev/publisher/internal/credentials"
 	"github.com/posit-dev/publisher/internal/events"
 	"github.com/posit-dev/publisher/internal/services/api"
 	"github.com/posit-dev/publisher/internal/util"
-	"github.com/r3labs/sse/v2"
 )
 
 type UICmd struct {
@@ -20,14 +21,8 @@ type UICmd struct {
 
 func (cmd *UICmd) Run(args *cli_types.CommonArgs, ctx *cli_types.CLIContext) error {
 	ctx.Logger.Info("Starting PublishUICmd.Run")
-	eventServer := sse.New()
-	ctx.Logger.Info("created event server")
-	eventServer.CreateStream("messages")
-	ctx.Logger.Info("created event stream")
 
-	emitter := events.NewSSEEmitter(eventServer)
-	log := events.NewLoggerWithSSE(args.Verbose, emitter)
-	ctx.Logger.Info("created SSE logger")
+	log := events.NewCLILogger(args.Verbose, os.Stderr)
 
 	absPath, err := cmd.Path.Abs()
 	if err != nil {
@@ -36,25 +31,18 @@ func (cmd *UICmd) Run(args *cli_types.CommonArgs, ctx *cli_types.CLIContext) err
 
 	credentials.UseKeychain = cmd.UseKeychain
 
-	// We need to create these only after the credentials.UseKeychain setting has been resolved.
-	// This is because NewCredentialsService will use the value of UseKeychain to determine
-	// whether to use the keychain or not.
 	accountList, err := accounts.NewAccountList(ctx.Fs, ctx.Logger)
 	if err != nil {
 		return err
 	}
 
-	// Auto-initialize if needed. This will be replaced by an API call from the UI
-	// for better error handling and startup performance.
 	svc := api.NewService(
 		"/",
 		cmd.Listen,
 		true,
 		absPath,
 		accountList,
-		log,
-		eventServer,
-		emitter)
+		log)
 	ctx.Logger.Info("created UI service")
 	return svc.Run()
 }
