+++ extensions/vscode/webviews/homeView/package.json
@@ -9,7 +9,6 @@
   },
   "dependencies": {
     "axios": "^1.15.0",
-    "eventsource": "^4.1.0",
     "pinia": "^3.0.4",
     "vue": "^3.5.32",
     "vue-virtual-scroller": "^2.0.0-beta.8"
