+++ internal/events/cli_logger_test.go
@@ -19,7 +19,7 @@ func TestCLILoggerSuite(t *testing.T) {
 	suite.Run(t, new(CLILoggerSuite))
 }
 
-func (s *LoggerSuite) TestNewCLILogger() {
+func (s *CLILoggerSuite) TestNewCLILogger() {
 	log := NewCLILogger(0, os.Stderr)
 	s.IsType(log.Handler(), &slog.TextHandler{})
 }
