--- internal/events/events_test.go
@@ -1,24 +1,0 @@
-package events
-
-// Copyright (C) 2023 by Posit Software, PBC.
-
-import (
-	"testing"
-
-	"github.com/posit-dev/publisher/internal/util/utiltest"
-	"github.com/stretchr/testify/suite"
-)
-
-type EventsSuite struct {
-	utiltest.Suite
-}
-
-func TestEventsSuite(t *testing.T) {
-	suite.Run(t, new(EventsSuite))
-}
-
-func (s *EventsSuite) TestEventTypeOf() {
-	expected := "agent/log"
-	actual := EventTypeOf(AgentOp, LogPhase)
-	s.Equal(expected, actual)
-}
