--- internal/events/sse_handler.go
@@ -1,80 +1,0 @@
-package events
-
-// Copyright (C) 2023 by Posit Software, PBC.
-
-import (
-	"context"
-	"log/slog"
-	"slices"
-
-	"github.com/posit-dev/publisher/internal/logging"
-)
-
-// SSEHandler emits an event for every log message.
-type SSEHandler struct {
-	emitter *SSEEmitter
-	attrs   []slog.Attr
-}
-
-var _ slog.Handler = &SSEHandler{}
-
-func NewSSEHandler(emitter *SSEEmitter) *SSEHandler {
-	return &SSEHandler{
-		emitter: emitter,
-	}
-}
-
-func (h *SSEHandler) Enabled(ctx context.Context, level slog.Level) bool {
-	return true
-}
-
-func (h *SSEHandler) recordToEvent(rec slog.Record) *Event {
-	event := &Event{
-		Time: rec.Time,
-		Data: make(EventData),
-	}
-	event.Data["Message"] = rec.Message
-	event.Data["Level"] = rec.Level.String()
-
-	// Convert the logged attributes into a map in the event Data.
-	// Also recognize the "op" and "phase" attributes as the context
-	// to use in the event "Type" field, e.g.
-	// log.Info("a message", "op", "publish/restore")
-	// will create an SSE event with Type: "publish/restore/log".
-	op := AgentOp
-
-	handleAttr := func(attr slog.Attr) bool {
-		switch attr.Key {
-		case logging.LogKeyOp:
-			op = Operation(attr.Value.String())
-		case "": // skip empty attrs
-		default:
-			event.Data[attr.Key] = attr.Value.Any()
-		}
-		return true
-	}
-	// First handle any attributes attached to this handler
-	for _, attr := range h.attrs {
-		handleAttr(attr)
-	}
-	// Then the ones from this specific message.
-	rec.Attrs(handleAttr)
-	event.Type = EventTypeOf(op, LogPhase)
-	return event
-}
-
-func (h *SSEHandler) Handle(ctx context.Context, rec slog.Record) error {
-	event := h.recordToEvent(rec)
-	return h.emitter.Emit(event)
-}
-
-func (h *SSEHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
-	newH := *h
-	newH.attrs = append(slices.Clip(newH.attrs), attrs...)
-	return &newH
-}
-
-func (h *SSEHandler) WithGroup(string) slog.Handler {
-	// This handler doesn't support groups.
-	return h
-}
